fix compiler warnings
many of those are due to conversion between signed and unsigned integer types.

+++ include/NeoFOAM/fields/domainField.hpp
@@ -36,7 +36,7 @@ public:
         : exec_(exec), internalField_(exec, 0), boundaryFields_(exec, 0, 0)
     {}
 
-    DomainField(const Executor& exec, int nCells, int nBoundaryFaces, int nBoundaries)
+    DomainField(const Executor& exec, size_t nCells, size_t nBoundaryFaces, size_t nBoundaries)
         : exec_(exec), internalField_(exec, nCells),
           boundaryFields_(exec, nBoundaryFaces, nBoundaries)
     {}
